local-file-inclusion/requirements.txt
Flask
pydantic
python_magic
